Move unit tests into dedicated directory in preparation for integration testing. Remove example file in favor of README.
